refactor(core): rename tools/auth.ts to tools/apply-auth.ts
tools/auth.ts only applies an existing HarnessConfig.auth to a
Playwright BrowserContext. The shorter name collided cognitively with
the other 8 auth-related files in the same directory (auth-detector,
auth-explorer, auth-surface-analyzer, auth-block-gap, oauth-providers,
user-providers). Renaming to apply-auth.ts spells out the function's
single responsibility.

No behavior change. No public API surface change — createAuthenticatedContext
keeps its name and is not exported from src/index.ts. The lone import
in playwright-explorer.ts is updated. Wider consolidation into a
tools/auth/ subfolder is deferred to a follow-up chore branch to keep
the storage-honesty PR readable.
