+++ packages/core/src/tools/apply-auth.ts
